+++ src/gui/GameGui.h
@@ -9,7 +9,7 @@ class GameGui : public RootGui {
 public:
 	GameGui(sf::Vector2f windowSize);
 	void setup(TurnManager &turnManager, sf::Font &font, unsigned int fontSize, sf::Vector2f childrenScale,
-		sf::Color backgroundColour, sf::Color hoverBackgroundColour, sf::Color borderColour, int borderThickness);
+		sf::Color backgroundColour, sf::Color hoverBackgroundColour, sf::Color borderColour, float borderThickness);
 	virtual void update(Input &input) override;
 
 	ConsoleGui *getConsoleGui();
